src/main/webapp/app/entities/quan-he-master/delete/quan-he-master-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { provideHttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { QuanHeMasterService } from '../service/quan-he-master.service';

import { QuanHeMasterDeleteDialogComponent } from './quan-he-master-delete-dialog.component';

describe('QuanHeMaster Management Delete Component', () => {
  let comp: QuanHeMasterDeleteDialogComponent;
  let fixture: ComponentFixture<QuanHeMasterDeleteDialogComponent>;
  let service: QuanHeMasterService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [QuanHeMasterDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(QuanHeMasterDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(QuanHeMasterDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(QuanHeMasterService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
